requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiomysql==0.2.0
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncclick==8.1.3.4
attrs==23.1.0
bcrypt==4.0.1
black==23.11.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
databases==0.8.0
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fastapi==0.104.1
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
idna==3.4
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.11.2
platformdirs==4.0.0
pydantic==1.10.11
pydantic_core==2.10.1
PyJWT==2.8.0
PyMySQL==1.1.0
python-decouple==3.8
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==1.4.50
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
yarl==1.9.2
